# src/main/resources/api-spec/product-api-spec.yml
openapi: 3.0.0
info:
  title: Product API
  description: API for managing products in the marketplace
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local server

paths:
  /products:
    get:
      summary: Retrieve all products
      description: Get a list of all products available in the marketplace.
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      summary: Create a new product
      description: Add a new product to the marketplace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '201':
          description: Product successfully created
        '400':
          description: Invalid request data

  /products/{id}:
    get:
      summary: Retrieve a product by ID
      description: Get details of a specific product using its unique identifier.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      summary: Update a product by ID
      description: Update the details of an existing product by its unique identifier.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '200':
          description: Product successfully updated
        '404':
          description: Product not found
    delete:
      summary: Delete a product by ID
      description: Remove a product from the marketplace using its unique identifier.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product successfully deleted
        '404':
          description: Product not found

components:
  schemas:
    ProductDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the product
        name:
          type: string
          description: Name of the product
        description:
          type: string
          description: Description of the product
        price:
          type: number
          format: float
          description: Price of the product
        origin:
          type: string
          description: Country of origin of the product
        category:
          type: string
          description: Category name to which the product belongs
      required:
        - name
        - price
        - origin